First batch of Placeholders PR comments
