+++ src/Pages/AddNewInvoice/AddNewInvoice.js
@@ -1,5 +1,6 @@
-import InvoiceForm from "../InvoiceForm";
+import React from "react";
+import InvoiceForm from "../../Components/InvoiceForm/InvoiceForm";
 
-const AddNewInvoice = () =>  <InvoiceForm />
+const AddNewInvoice = () => <InvoiceForm />;
 
-export default AddNewInvoice
+export default AddNewInvoice;
