+++ src/Pages/EditInvoice/EditInvoice.js
@@ -1,10 +1,7 @@
 import React from "react";
-import { useParams } from "react-router-dom";
-import mockedData from "../../mockedData.json";
 
-import InvoiceForm from "../InvoiceForm";
+import InvoiceForm from "../../Components/InvoiceForm/InvoiceForm";
 
 export default function EditInvoice() {
-  const params = useParams();
-  <InvoiceForm preDefinedId={params.invoiceId} preDefinedFields={mockedData[params.invoiceId]} />
+  return <InvoiceForm />;
 }
